RDKEMW-6872: Reduce Max Sleep for Previous Report and Fix Crash i
Reason for change: A crash was observed when publishReportUploadStatus was called for requests that had already been processed. This was due to the handler not being set to NULL after handling a request, leading to double invocation.
Test Procedure: There should not be crash in publishReportUploadStatus  function call and regression issues
Risks: Medium
Priority: P1